# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Normalized MOEX bond and Bank of Russia rate data.",
        "title": "MOEX Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check gateway health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.HealthResponse"
                        }
                    }
                }
            }
        },
        "/v1/bonds": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bonds"
                ],
                "summary": "List MOEX bonds",
                "parameters": [
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "default": 40,
                        "description": "Maximum number of bonds",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/moex.Bond"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/bonds/{isin}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bonds"
                ],
                "summary": "Get a MOEX bond",
                "parameters": [
                    {
                        "maxLength": 12,
                        "minLength": 12,
                        "type": "string",
                        "description": "ISIN",
                        "name": "isin",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/moex.Bond"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/cbr/rates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cbr"
                ],
                "summary": "Get Bank of Russia rate data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cbr.RateSnapshot"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cbr.Direction": {
            "type": "string",
            "enum": [
                "up",
                "flat",
                "down",
                "unknown"
            ],
            "x-enum-varnames": [
                "DirectionUp",
                "DirectionFlat",
                "DirectionDown",
                "DirectionUnknown"
            ]
        },
        "cbr.RateForecast": {
            "type": "object",
            "properties": {
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "midpoint": {
                    "type": "number"
                },
                "published_at": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "cbr.RatePoint": {
            "type": "object",
            "properties": {
                "effective_date": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                }
            }
        },
        "cbr.RateSnapshot": {
            "type": "object",
            "properties": {
                "current_rate": {
                    "type": "number"
                },
                "direction": {
                    "$ref": "#/definitions/cbr.Direction"
                },
                "fetched_at": {
                    "type": "string"
                },
                "forecast": {
                    "$ref": "#/definitions/cbr.RateForecast"
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cbr.RatePoint"
                    }
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "httpserver.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "httpserver.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "moex.Bond": {
            "type": "object",
            "properties": {
                "accrued_interest": {
                    "type": "number"
                },
                "bond_subtype": {
                    "type": "string"
                },
                "bond_type": {
                    "type": "string"
                },
                "cashflow_schedule": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/moex.Cashflow"
                    }
                },
                "coupon_calendar": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/moex.Coupon"
                    }
                },
                "coupon_frequency": {
                    "type": "integer"
                },
                "coupon_percent": {
                    "type": "number"
                },
                "coupon_type": {
                    "type": "string"
                },
                "coupon_value": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "days_to_redemption": {
                    "type": "integer"
                },
                "duration": {
                    "type": "number"
                },
                "evening_session": {
                    "type": "boolean"
                },
                "face_unit": {
                    "type": "string"
                },
                "face_value": {
                    "type": "number"
                },
                "has_default": {
                    "type": "boolean"
                },
                "has_prospectus": {
                    "type": "boolean"
                },
                "has_technical_default": {
                    "type": "boolean"
                },
                "initial_face_value": {
                    "type": "number"
                },
                "instrument_group": {
                    "type": "string"
                },
                "is_qualified_investors": {
                    "type": "boolean"
                },
                "isin": {
                    "type": "string"
                },
                "issue_date": {
                    "type": "string"
                },
                "issue_size": {
                    "type": "number"
                },
                "issuer_id": {
                    "type": "integer"
                },
                "latin_name": {
                    "type": "string"
                },
                "list_level": {
                    "type": "integer"
                },
                "lot_size": {
                    "type": "integer"
                },
                "market_data_as_of": {
                    "type": "string"
                },
                "maturity_date": {
                    "type": "string"
                },
                "morning_session": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "next_coupon_date": {
                    "type": "string"
                },
                "num_trades": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "reg_number": {
                    "type": "string"
                },
                "registry_date": {
                    "type": "string"
                },
                "secid": {
                    "type": "string"
                },
                "short_name": {
                    "type": "string"
                },
                "start_date_moex": {
                    "type": "string"
                },
                "ticker": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "type_name": {
                    "type": "string"
                },
                "value_today": {
                    "type": "number"
                },
                "weekend_session": {
                    "type": "boolean"
                },
                "yield_to_maturity": {
                    "type": "number"
                }
            }
        },
        "moex.Cashflow": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "taxable": {
                    "type": "boolean"
                }
            }
        },
        "moex.Coupon": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}